Deleting local.properties file so it's ignored
local.properties file says to exclude this from the commit (by adding it to the gitignore), and it's already been committed so now we need to remove the file from the repo and simply re-add it on our own IDE's. It is different for each user, which is why it can't be added to the repository, since it will be specific to our own environments.